@@ -4,27 +4,25 @@
   "description": "Using webpack with yFiles for HTML",
   "main": "webpack-demo.js",
   "author": "yfileshtml@yworks.com",
-  "license": "https://www.yworks.com/en/products_yfiles_sla.html",
+  "license": "https://www.yworks.com/products/yfiles-for-html/sla",
   "private": true,
   "dependencies": {
     "color": "3.0.0"
   },
   "scripts": {
-    "production": "grunt && webpack --config webpack.prod.js",
-    "build": "webpack --config webpack.dev.js",
-    "start": "webpack --config webpack.dev.js && webpack-dev-server --config webpack.dev.js --open"
+    "production": "grunt && webpack --mode production",
+    "build": "webpack --mode development",
+    "start": "webpack --mode development && webpack-dev-server --mode development --open"
   },
   "devDependencies": {
-    "chunk-manifest-webpack-plugin": "1.1.2",
-    "grunt": "1.0.1",
+    "grunt": "1.0.3",
     "grunt-contrib-clean": "1.1.0",
     "grunt-yfiles-deployment": "../../../deployment/grunt-yfiles-deployment",
-    "html-webpack-harddisk-plugin": "0.1.0",
-    "html-webpack-plugin": "2.30.1",
-    "uglifyjs-webpack-plugin": "1.1.8",
-    "webpack": "3.10.0",
-    "webpack-chunk-hash": "0.5.0",
-    "webpack-dev-server": "2.11.1",
-    "webpack-merge": "4.1.2"
+    "html-webpack-harddisk-plugin": "0.2.0",
+    "html-webpack-plugin": "3.2.0",
+    "webpack": "4.16.5",
+    "webpack-cli": "3.1.0",
+    "webpack-dev-server": "3.1.5",
+    "webpack-merge": "4.1.4"
   }
 }
