@@ -8,24 +8,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "6.0.2",
-    "@angular/common": "6.0.2",
-    "@angular/compiler": "6.0.2",
-    "@angular/core": "6.0.2",
-    "@angular/forms": "6.0.2",
-    "@angular/http": "6.0.2",
-    "@angular/platform-browser": "6.0.2",
-    "@angular/platform-browser-dynamic": "6.0.2",
-    "@angular/router": "6.0.2",
-    "core-js": "2.4.1",
-    "rxjs": "6.1.0",
+    "@angular/animations": "6.1.3",
+    "@angular/common": "6.1.3",
+    "@angular/compiler": "6.1.3",
+    "@angular/core": "6.1.3",
+    "@angular/forms": "6.1.3",
+    "@angular/http": "6.1.3",
+    "@angular/platform-browser": "6.1.3",
+    "@angular/platform-browser-dynamic": "6.1.3",
+    "@angular/router": "6.1.3",
+    "core-js": "2.5.7",
+    "rxjs": "6.2.2",
     "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^6.0.1",
-    "@angular/compiler-cli": "6.0.2",
-    "@angular/language-service": "6.0.2",
-    "typescript": "2.7.2",
-    "@angular-devkit/build-angular": "~0.6.1"
+    "@angular/cli": "6.1.4",
+    "@angular/compiler-cli": "6.1.3",
+    "@angular/language-service": "6.1.3",
+    "typescript": "2.9.2",
+    "@angular-devkit/build-angular": "0.7.4"
   }
 }
