@@ -3,19 +3,20 @@
   "version": "2.0.0",
   "private": true,
   "dependencies": {
-    "react": "15.6.2",
-    "react-dom": "15.6.2"
+    "react": "16.4.2",
+    "react-dom": "16.4.2"
   },
   "scripts": {
     "babel": "babel src -d build",
-    "webpack": "webpack",
-    "build": "npm run babel && npm run webpack",
+    "webpack": "webpack --mode=production",
+    "build": "npm run babel && npm run webpack --mode=production",
     "build:watch": "watch \"npm run build\" src"
   },
   "devDependencies": {
     "babel-cli": "6.26.0",
     "babel-preset-react": "6.24.1",
     "watch": "1.0.2",
-    "webpack": "3.10.0"
+    "webpack": "4.17.1",
+    "webpack-cli": "3.1.0"
   }
 }
