@@ -3,7 +3,7 @@
   "version": "2.0.0",
   "dependencies": {},
   "devDependencies": {
-    "typescript": "2.6.2"
+    "typescript": "3.0.1"
   },
   "scripts": {
     "postinstall": "npm run build",
