@@ -1,12 +1,11 @@
 {
   "name": "TypeScript-Demo",
   "version": "2.0.1",
-  "dependencies": {
-    "ts-loader": "3.3.1",
-    "webpack": "3.10.0"
-  },
   "devDependencies": {
-    "typescript": "2.6.2"
+    "ts-loader": "4.5.0",
+    "typescript": "3.0.1",
+    "webpack": "4.17.1",
+    "webpack-cli": "3.1.0"
   },
   "scripts": {
     "postinstall": "npm run build",
