@@ -763,7 +763,7 @@ function getDemoData() {
       thumbnailPath: 'resources/image/mazerouting.png',
       sourcePath: 'layout/mazerouting/MazeRoutingDemo.js',
       tags: ['layout', 'router'],
-      keywords: ['edgerouter', 'polyline-router']
+      keywords: ['edgerouter', 'polylinerouter']
     },
     {
       id: 'interactive-map',
@@ -990,12 +990,12 @@ function getDemoData() {
       name: 'Clustering Algorithms',
       demoPath: 'analysis/clustering/index.html',
       summary:
-        'Showcases a selection of clustering algorithms such as edge-betweenness, k-means, hierarchical and biconnected components clustering.',
+        'Showcases a selection of clustering algorithms such as edge betweenness, k-means, hierarchical and biconnected components clustering.',
       category: 'analysis',
       thumbnailPath: 'resources/image/clustering.png',
       sourcePath: 'analysis/clustering/ClusteringDemo.js',
       tags: ['algorithms'],
-      keywords: ['k-means', 'hierarchical', 'voronoi', 'dendrogram']
+      keywords: ['kmeans', 'hierarchical', 'voronoi', 'dendrogram']
     },
     {
       id: 'network-flows',
@@ -1364,7 +1364,7 @@ function getDemoData() {
       name: 'NodeJS',
       demoPath: 'loading/nodejs/index.html',
       summary:
-        " Shows how to run a yFiles layout algorithm in a <a href='http://nodejs.org/' target='_blank'>Node.js&reg;</a> environment.",
+        " Shows how to run a yFiles layout algorithm in a <a href='https://nodejs.org/' target='_blank'>Node.js&reg;</a> environment.",
       category: 'loading',
       thumbnailPath: 'resources/image/nodejs.png',
       sourcePath: 'loading/nodejs/NodeJsDemo.js',
@@ -1475,7 +1475,7 @@ function getDemoData() {
       name: 'HTML Label',
       demoPath: 'style/htmllabel/index.html',
       summary: 'Shows how HTML can be used in label text with a custom label style.',
-      category: 'styles',
+      category: 'style',
       thumbnailPath: 'resources/image/htmllabel.png',
       sourcePath: 'style/htmllabel/HtmlLabelDemo.js',
       tags: ['styles', 'labels'],
@@ -1601,7 +1601,7 @@ function getDemoData() {
       name: 'Browserify',
       demoPath: 'loading/browserify/README.html',
       summary:
-        "Shows how to bundle the yFiles library in a <a href='http://browserify.org/'>Browserify</a> project.",
+        "Shows how to bundle the yFiles library in a <a href='https://browserify.org/'>Browserify</a> project.",
       category: 'loading',
       thumbnailPath: 'resources/image/browserify.png',
       sourcePath: 'loading/browserify/browserify-demo.js',
@@ -1674,7 +1674,7 @@ function getDemoData() {
       category: 'view',
       thumbnailPath: 'resources/image/renderingorder.png',
       sourcePath: 'view/renderingorder/RenderingOrderDemo.js',
-      tags: ['rendering', 'z-order', 'nesting'],
+      tags: ['rendering', 'zorder', 'nesting'],
       keywords: ['hierarchicnesting', 'ports', 'labels', 'groups']
     },
     {
@@ -1873,7 +1873,7 @@ function getDemoData() {
       demoPath: 'style/sparkline/index.html',
       summary:
         'Shows how to integrate visualizations generated by a third-party library with a custom node style.',
-      category: 'styles',
+      category: 'style',
       thumbnailPath: 'resources/image/sparklines.png',
       sourcePath: 'style/sparkline/SparklineDemo.html',
       tags: ['styles'],
