@@ -32,6 +32,7 @@ const path = require('path')
 
 module.exports = {
   entry: './TypeScript2Demo.ts',
+  mode: 'development',
   output: {
     filename: './bundle.js'
   },
@@ -41,7 +42,7 @@ module.exports = {
     extensions: ['.webpack.js', '.web.js', '.ts', '.tsx', '.js']
   },
   module: {
-    loaders: [
+    rules: [
       // all files with a '.ts' or '.tsx' extension will be handled by 'ts-loader'
       {
         test: /\.tsx?$/,
